pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-model-augmentation-library"
version = "0.1.0"
description = "A lightweight image augmentation package"
authors = [
    { name = "Dylan Tran", email = "trandf@uci.edu" }
]
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "pillow"
]

[tool.setuptools]
py-modules = []

[project.urls]
"Homepage" = "https://github.com/dtro18/imageAugmentation"  